Another round of fixes
